Do location stuff from the universe object
Giving locations access to the universe as an attribute makes some
things too complicated. Plus it is a bit of code smell. If we call
methods from the universe and passing the location in, you have access
to both without needing to deal with weak refs or weird schema
workarounds.