Make 'quit' work again by taking advantage of the post-response hook
With the refactor, 'quit' stopped working because disconnects were
reworked to be handled per-backend. Quit could not be implemented until
there was a mark-and-sweep system to prevent premature disconnections.
With this, markings are respected when closing sockets after responses
are finished (heavy-lifting done by Reflex) and when determining whether
to display the prompt.