Pass the composite and an event object instead of arbitrary vars
This allows for more flexibility with the command system as more
features get added. For instance, another entity is merely another
attribute that doesn't require changing code in every single command.
Also, it allows for a space to provide convenience methods for the event
and the composite object to be used right in the command block.